+++ src/main/java/kitchenpos/eatinorders/domain/EatInOrderRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.eatinorders.domain;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.eatinordertables.domain.EatInOrderTable;
 
 public interface EatInOrderRepository {
     EatInOrder save(EatInOrder eatInOrder);
