+++ src/main/java/kitchenpos/eatinordertables/domain/EatInOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinordertables.domain;
 
 import java.util.List;
 import java.util.Optional;
