+++ src/test/java/kitchenpos/eatinorders/EatInOrderFixtures.java
@@ -6,18 +6,11 @@ import java.util.List;
 import java.util.Random;
 import java.util.UUID;
 import kitchenpos.eatinorders.domain.*;
+import kitchenpos.eatinordertables.domain.EatInOrderTable;
 
 public class EatInOrderFixtures {
 
     public static final UUID INVALID_ID = new UUID(0L, 0L);
-
-    public static EatInOrderTable eatInOrderTable() {
-        return eatInOrderTable(false, 0);
-    }
-
-    public static EatInOrderTable eatInOrderTable(final boolean occupied, final int numberOfGuests) {
-        return new EatInOrderTable("1번", numberOfGuests, occupied);
-    }
 
     public static EatInOrder eatInOrder(final EatInOrderStatus status, final EatInOrderTable eatInOrderTable) {
         return new EatInOrder(
