+++ src/test/java/kitchenpos/eatinordertables/EatInOrderTableFixtures.java
@@ -1,0 +1,14 @@
+package kitchenpos.eatinordertables;
+
+import kitchenpos.eatinordertables.domain.EatInOrderTable;
+
+public class EatInOrderTableFixtures {
+
+    public static EatInOrderTable eatInOrderTable() {
+        return eatInOrderTable(false, 0);
+    }
+
+    public static EatInOrderTable eatInOrderTable(final boolean occupied, final int numberOfGuests) {
+        return new EatInOrderTable("1번", numberOfGuests, occupied);
+    }
+}
